question_4_python_llm/requirements.txt
pandas
langchain
langchain-openai
python-dotenv
pydantic